requirements.txt
fastapi
uvicorn[standard]
resend
python-dotenv
pydantic
retell-sdk
ruff